wip the mesh and prism class working right normals for basic effect lighting ect...
Fixed the normals in the mesh fixed the winding in the primitives
Added tangents to the mesh and normals to the primitives.
Straightened out the usage for basic effects so its not as wasteful.
Added a setup method before drawing for multiple draws of the same type of primitive.  Still more stuff need to add a method to the mesh to take a texture for the height data. ect more clean up all over.